Convert [ParamsSource] to [ArgumentsSource]
