+++ tests/Benchmarks/Utf8String/CodePointEnumeration.cs
@@ -3,7 +3,6 @@
 
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Code;
-using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Text;
@@ -13,25 +12,17 @@ namespace Benchmarks.Utf8String
 {
     public partial class Utf8String
     {
-        [ParamsSource(nameof(GetConstructFromStringParameters))]
-        public string ConstructFromStringData;
-
-        public IEnumerable<IParam> GetConstructFromStringParameters()
-        {
-            yield return new ConstructFromStringParameter(5, 32, 126, "Short ASCII string");
-            yield return new ConstructFromStringParameter(5, 32, 0xD7FF, "Short string");
-            yield return new ConstructFromStringParameter(50000, 32, 126, "Long ASCII string");
-            yield return new ConstructFromStringParameter(50000, 32, 0xD7FF, "Long string");
-        }
-
         [Benchmark]
-        public Utf8StringRealType ConstructFromString()
+        [ArgumentsSource(nameof(GetEnumerateCodePointsParameters))]
+        public uint EnumerateCodePoints(Utf8StringRealType value)
         {
-            return new Utf8StringRealType(ConstructFromStringData);
+            uint lastValue = default;
+            foreach (var codePoint in value)
+            {
+                lastValue = codePoint;
+            }
+            return lastValue;
         }
-
-        [ParamsSource(nameof(GetEnumerateCodePointsParameters))]
-        public Utf8StringRealType EnumerateCodePointsData;
 
         public IEnumerable<IParam> GetEnumerateCodePointsParameters()
         {
@@ -40,56 +31,7 @@ namespace Benchmarks.Utf8String
             yield return new EnumerateCodePointsParameter(50000, 32, 126, "Long ASCII string");
             yield return new EnumerateCodePointsParameter(50000, 32, 0xD7FF, "Long string");
         }
-
-        [Benchmark]
-        public uint EnumerateCodePoints()
-        {
-            uint lastValue = default;
-            foreach (var codePoint in EnumerateCodePointsData)
-            {
-                lastValue = codePoint;
-            }
-            return lastValue;
-        }
-
-        private static string GetRandomString(int length, int minCodePoint, int maxCodePoint)
-        {
-            Random r = new Random(42);
-            StringBuilder sb = new StringBuilder(length);
-            while (length-- != 0)
-            {
-                sb.Append((char)r.Next(minCodePoint, maxCodePoint));
-            }
-            return sb.ToString();
-        }
-
-        public class ConstructFromStringParameter : IParam
-        {
-            public ConstructFromStringParameter(int length, int minCodePoint, int maxCodePoint, string description)
-            {
-                DisplayText = description;
-                Value = GetRandomString(length, minCodePoint, maxCodePoint);
-            }
-
-            public string DisplayText { get; }
-
-            public object Value { get; }
-
-            public string ToSourceCode()
-            {
-                string valueAsString = (string)Value;
-
-                StringBuilder sb = new StringBuilder("\"");
-                foreach (char ch in valueAsString)
-                {
-                    sb.AppendFormat(CultureInfo.InvariantCulture, "\\u{0:X4}", (uint)ch);
-                }
-                sb.Append("\"");
-
-                return sb.ToString();
-            }
-        }
-
+        
         public class EnumerateCodePointsParameter : IParam
         {
             private string _value;
@@ -116,6 +58,5 @@ namespace Benchmarks.Utf8String
                 return sb.ToString();
             }
         }
-
     }
 }
